chore: Refactor theme imports and clean up configuration files
- Update theme import paths in multiple components to reflect new structure
- Remove unused Babel plugins from babel.config.js for cleaner configuration
- Delete obsolete theme.ts file to streamline project organization
- Update package-lock.json to reflect changes in dependencies
