+++ app/App.tsx
@@ -1,6 +1,6 @@
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View } from 'react-native';
-import { theme } from './app/theme';
+import { theme } from './theme';
 
 export default function App() {
   return (
