+++ app/app/index.tsx
@@ -6,7 +6,7 @@ import {
   TouchableOpacity,
   StyleSheet,
 } from 'react-native';
-import { theme } from './theme';
+import { theme } from '../theme';
 
 export default function TodosScreen() {
   return (
