+++ app/app/settings.tsx
@@ -7,7 +7,7 @@ import {
   StyleSheet,
   Switch,
 } from 'react-native';
-import { theme } from './theme';
+import { theme } from '../theme';
 
 export default function SettingsScreen() {
   return (
