+++ app/babel.config.js
@@ -2,8 +2,5 @@ module.exports = function (api) {
   api.cache(true);
   return {
     presets: ['babel-preset-expo'],
-    plugins: [
-      'expo-router/babel'
-    ],
   };
-}; 
+};
